Extract `runWatch` into separate module
Also remove default values from `runWatch` they are already set by
`runMocha`.
