+++ lib/cli/run-helpers.js
@@ -12,11 +12,12 @@ const path = require('path');
 const ansi = require('ansi-colors');
 const debug = require('debug')('mocha:cli:run:helpers');
 const minimatch = require('minimatch');
-const Context = require('../context');
-const Mocha = require('../mocha');
 const utils = require('../utils');
+const watchRun = require('./watch-run');
 
 const cwd = (exports.cwd = process.cwd());
+
+exports.watchRun = watchRun;
 
 /**
  * Exits Mocha when tests + code under test has finished execution (default)
@@ -63,32 +64,6 @@ const exitMocha = code => {
   });
 
   done();
-};
-
-/**
- * Hide the cursor.
- * @ignore
- * @private
- */
-const hideCursor = () => {
-  process.stdout.write('\u001b[?25l');
-};
-
-/**
- * Show the cursor.
- * @ignore
- * @private
- */
-const showCursor = () => {
-  process.stdout.write('\u001b[?25h');
-};
-
-/**
- * Stop cursor business
- * @private
- */
-const stop = () => {
-  process.stdout.write('\u001b[2K');
 };
 
 /**
@@ -208,76 +183,6 @@ exports.singleRun = (mocha, {files = [], exit = false} = {}) => {
 };
 
 /**
- * Run Mocha in "watch" mode
- * @param {Mocha} mocha - Mocha instance
- * @param {Object} [opts] - Options
- * @param {string[]} [opts.extension] - List of extensions to watch
- * @param {string|RegExp} [opts.grep] - Grep for test titles
- * @param {string} [opts.ui=bdd] - User interface
- * @param {string[]} [files] - Array of test files
- * @private
- */
-exports.watchRun = (
-  mocha,
-  {extension = [], grep = '', ui = 'bdd', files = []} = {}
-) => {
-  let runner;
-
-  console.log();
-  hideCursor();
-  process.on('SIGINT', () => {
-    showCursor();
-    console.log('\n');
-    process.exit(130);
-  });
-
-  const watchFiles = utils.files(cwd, extension);
-  let runAgain = false;
-
-  const loadAndRun = () => {
-    try {
-      mocha.files = files;
-      runAgain = false;
-      runner = mocha.run(() => {
-        runner = null;
-        if (runAgain) {
-          rerun();
-        }
-      });
-    } catch (e) {
-      console.log(e.stack);
-    }
-  };
-
-  const purge = () => {
-    watchFiles.forEach(Mocha.unloadFile);
-  };
-
-  loadAndRun();
-
-  const rerun = () => {
-    purge();
-    stop();
-    if (!grep) {
-      mocha.grep(null);
-    }
-    mocha.suite = mocha.suite.clone();
-    mocha.suite.ctx = new Context();
-    mocha.ui(ui);
-    loadAndRun();
-  };
-
-  utils.watch(watchFiles, () => {
-    runAgain = true;
-    if (runner) {
-      runner.abort();
-    } else {
-      rerun();
-    }
-  });
-};
-
-/**
  * Actually run tests
  * @param {Mocha} mocha - Mocha instance
  * @param {Object} [opts] - Options
@@ -295,7 +200,7 @@ exports.runMocha = (
   files = []
 ) => {
   if (watch) {
-    exports.watchRun(mocha, {extension, grep, ui, files});
+    watchRun(mocha, {extension, grep, ui, files});
   } else {
     exports.singleRun(mocha, {files, exit});
   }
